requirements.txt
anyio==3.3.0
appnope==0.1.2
argon2-cffi==20.1.0
attrs==21.2.0
Automat==20.2.0
awscli==1.19.72
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
bleach==4.0.0
botocore==1.20.72
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.6
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==3.4.8
cssselect==1.1.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
docutils==0.15.2
entrypoints==0.3
filelock==3.0.12
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
itemadapter==0.4.0
itemloaders==1.0.4
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.10.2
jupyterlab==3.1.7
jupyterlab-pygments==0.1.2
jupyterlab-server==2.7.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclassic==0.3.1
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
packaging==21.0
pandocfilters==1.4.3
parsel==1.6.0
parso==0.8.2
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.3.0
priority==1.3.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
Protego==0.1.16
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.10.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.2.1
queuelib==1.6.2
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.7.2
s3transfer==0.4.2
Scrapy==2.5.0
selenium==3.141.0
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
stevedore==3.4.0
terminado==0.11.0
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
Twisted==21.7.0
typing-extensions==3.10.0.2
urllib3==1.25.11
virtualenv==20.7.2
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
zope.interface==5.4.0
